refactor: merge e2e tests into single file with fixtures
Consolidate all e2e tests into a single `e2e.rs` file and ensure
all tests use fixture files instead of inline source strings.

Changes:
- Merged `e2e_hello.rs` and `e2e_effect_import.rs` into `e2e.rs`
- Created `fixtures/effect_import_aliasing.wado`
- Created `fixtures/multiple_effect_imports.wado`
- All 5 e2e tests now use `include_str!("fixtures/...")` pattern
- Removed inline source code strings from tests

All tests pass successfully (5 e2e tests + 13 unit tests).
