+++ wado-compiler/tests/fixtures/effect_import_aliasing.wado
@@ -1,0 +1,8 @@
+// Test for Effect import with aliasing
+
+use core::cli::{println, Stdout};
+
+fn main() with Stdout {
+    println("Aliased import works!");
+    println("Direct import also works!");
+}
