+++ wado-compiler/tests/fixtures/multiple_effect_imports.wado
@@ -1,0 +1,7 @@
+// Test for importing from multiple effects
+
+use core::cli::{println, Stdout};
+
+fn main() with Stdout {
+    println("Testing multiple imports");
+}
